added auto-reconnect flag to listen.
The function signature is now: `cumin.listen(queueName, handler, autoReconnect)`. If `autoReconnect` is a truthy value,
then handlers are attached to both the block and the non-blocking client, so when a Redis connection drops, the process
won't exit. In addition, when the blocking client reconnects, it restarts listening for items from the queue.

I haven't added any tests for this functionality (I didn't want to expose the redis-clients just for testing), but I
would be happy to add them if you have a suggested way of doing it.